PWN-9313 - Move mapping to converter
