+++ ethereumkit/src/main/java/org/p2p/ethereumkit/external/model/EthTokenConverter.kt
@@ -1,8 +1,11 @@
 package org.p2p.ethereumkit.external.model
 
 import java.math.BigDecimal
+import java.math.BigInteger
 import org.p2p.core.token.Token
 import org.p2p.core.utils.fromLamports
+import org.p2p.core.wrapper.eth.EthAddress
+import org.p2p.token.service.model.TokenServiceMetadata
 
 object EthTokenConverter {
 
@@ -33,6 +36,24 @@ object EthTokenConverter {
             rate = metadata.price,
             isClaiming = isClaiming,
             latestActiveBundleId = bundleId
+        )
+    }
+
+    fun toEthTokenMetadata(
+        ethAddress: EthAddress,
+        metadata: TokenServiceMetadata,
+        tokenBalance: BigInteger
+    ): EthTokenMetadata {
+        val erc20Token = ERC20Tokens.findToken(ethAddress)
+        return EthTokenMetadata(
+            contractAddress = ethAddress,
+            mintAddress = erc20Token.mintAddress,
+            balance = tokenBalance,
+            decimals = metadata.decimals,
+            logoUrl = metadata.logoUrl,
+            tokenName = erc20Token.replaceTokenName.orEmpty(),
+            symbol = erc20Token.replaceTokenSymbol.orEmpty(),
+            price = null
         )
     }
 }
