+++ ethereumkit/src/main/java/org/p2p/ethereumkit/external/model/EthTokenMetadata.kt
@@ -2,8 +2,6 @@ package org.p2p.ethereumkit.external.model
 
 import java.math.BigDecimal
 import java.math.BigInteger
-import org.p2p.ethereumkit.external.api.alchemy.response.TokenBalanceResponse
-import org.p2p.ethereumkit.external.api.alchemy.response.TokenMetadataResponse
 import org.p2p.core.wrapper.eth.EthAddress
 
 data class EthTokenMetadata(
@@ -16,20 +14,3 @@ data class EthTokenMetadata(
     val symbol: String,
     var price: BigDecimal? = null
 )
-
-internal fun mapToTokenMetadata(
-    balanceResponse: TokenBalanceResponse,
-    metadata: TokenMetadataResponse,
-    erc20Token: ERC20Tokens
-): EthTokenMetadata {
-
-    return EthTokenMetadata(
-        contractAddress = balanceResponse.contractAddress,
-        mintAddress = erc20Token.mintAddress,
-        balance = balanceResponse.tokenBalance,
-        decimals = metadata.decimals,
-        logoUrl = erc20Token.tokenIconUrl ?: metadata.logoUrl.orEmpty(),
-        tokenName = erc20Token.replaceTokenName ?: metadata.tokenName.orEmpty(),
-        symbol = erc20Token.replaceTokenSymbol ?: metadata.symbol.orEmpty()
-    )
-}
